added HashBTC

+++ src/main/java/com/kurotkin/rate/RateRepository.java
@@ -2,5 +2,5 @@ package com.kurotkin.rate;
 
 import org.springframework.data.repository.CrudRepository;
 
-public interface RateRepository extends CrudRepository<Rate, Integer> {
+public interface RateRepository extends CrudRepository<Rate, Long> {
 }
